+++ src/main/java/com/yapp/artie/domain/user/deprecated/UserJpaEntityLoader.java
@@ -2,7 +2,7 @@ package com.yapp.artie.domain.user.deprecated;
 
 import com.yapp.artie.domain.user.adapter.out.persistence.UserJpaEntity;
 import com.yapp.artie.domain.user.adapter.out.persistence.UserRepository;
-import com.yapp.artie.domain.user.exception.UserNotFoundException;
+import com.yapp.artie.domain.user.application.port.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
