+++ src/test/java/com/yapp/artie/domain/user/application/service/GetUserServiceTest.java
@@ -8,7 +8,7 @@ import static org.mockito.BDDMockito.given;
 
 import com.yapp.artie.domain.user.application.port.out.LoadUserPort;
 import com.yapp.artie.domain.user.domain.User;
-import com.yapp.artie.domain.user.exception.UserNotFoundException;
+import com.yapp.artie.domain.user.application.port.exception.UserNotFoundException;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
